+++ app/src/main/java/com/sdzshn3/android/news247/ViewModel/TopHeadlinesViewModel.java
@@ -11,6 +11,7 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 import com.sdzshn3.android.news247.Fragments.TopHeadlinesFragment;
+import com.sdzshn3.android.news247.Repositories.TopHeadlinesRepository;
 import com.sdzshn3.android.news247.Retrofit.ApiService;
 import com.sdzshn3.android.news247.Retrofit.Article;
 import com.sdzshn3.android.news247.Retrofit.NewsModel;
@@ -19,53 +20,19 @@ import com.sdzshn3.android.news247.Retrofit.Client;
 import java.util.List;
 
 public class TopHeadlinesViewModel extends AndroidViewModel {
-    private static MutableLiveData<List<Article>> data = new MutableLiveData<>();
-    private static Call<NewsModel> call;
-    private ApiService apiService;
+
+    private TopHeadlinesRepository repository;
 
     public TopHeadlinesViewModel(Application application) {
         super(application);
-        apiService = Client.getApiService();
-        loadData();
-    }
-
-    private void loadData() {
-        call = apiService.getResponse(TopHeadlinesFragment.URL);
-        call.enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("TopHeadlinesViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        repository = new TopHeadlinesRepository();
     }
 
     public void refresh() {
-        call = apiService.getResponse(TopHeadlinesFragment.URL);
-        call.clone().enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("TopHeadlinesViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        repository.refresh();
     }
 
     public LiveData<List<Article>> getData() {
-        return data;
+        return repository.getData();
     }
 }
